requirements-7.txt
python-telegram-bot==21.6
psycopg2-binary
python-dotenv
fastapi
uvicorn[standard]
